backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
telethon==1.36.0
sqlalchemy==2.0.35
python-dotenv==1.0.1
websockets==13.1
python-multipart==0.0.12
aiofiles==24.1.0
pytz==2024.2
MetaTrader5>=5.0.4000
yfinance>=0.2.30
tvdatafeed>=2.1.0
pandas>=2.0.0
numpy>=1.24.0
